refactor(portal): update component folder naming (ceph)
