+++ packages/aurora/src/client/routes/_auth/projects/$projectId/storage/$provider/$storageType/index.tsx
@@ -3,7 +3,7 @@ import { z } from "zod"
 import { getServiceIndex } from "@/server/Authentication/helpers"
 import { ErrorBoundary } from "react-error-boundary"
 import { SwiftContainers } from "../../-components/Swift/Containers"
-import { CephBuckets } from "../../-components/Ceph/Containers"
+import { CephBuckets } from "../../-components/Ceph/Buckets"
 import { Trans, useLingui } from "@lingui/react/macro"
 import type { RouteInfo } from "@/client/routes/routeInfo"
 import { ContentHeader } from "@/client/components/ContentHeader/ContentHeader"
