+++ packages/aurora/src/client/routes/_auth/projects/$projectId/storage/-components/Ceph/Buckets/EnableVersioningModal.tsx
