+++ packages/aurora/src/client/routes/_auth/projects/$projectId/storage/-components/Ceph/Objects/ObjectBrowserView.tsx
@@ -18,8 +18,8 @@ import { SortSettings } from "@/client/components/ListToolbar/types"
 import { ObjectsTableView } from "./ObjectsTableView"
 import { ObjectsFileNavigation } from "./ObjectsFileNavigation"
 import { CreateFolderModal } from "./CreateFolderModal"
-import { EnableVersioningModal } from "../Containers/EnableVersioningModal"
-import { SuspendVersioningModal } from "../Containers/SuspendVersioningModal"
+import { EnableVersioningModal } from "../Buckets/EnableVersioningModal"
+import { SuspendVersioningModal } from "../Buckets/SuspendVersioningModal"
 import { useNavigate } from "@tanstack/react-router"
 import { Route } from "@/client/routes/_auth/projects/$projectId/storage/$provider/$storageType/$containerName/objects"
 import type { S3Object, S3FolderPrefix } from "@/server/Storage/types/ceph"
